Move bookmark tapped code
